% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corporaexplorer-deprecated.R,
%   R/run_corpus_explorer.R
\name{corporaexplorer-deprecated}
\alias{corporaexplorer-deprecated}
\alias{run_corpus_explorer}
\title{Deprecated functions in package \pkg{corporaexplorer}}
\usage{
run_corpus_explorer(...)
}
\arguments{
\item{...}{For \code{run_corpus_explorer},
arguments passed to \code{explore()}}
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future.
}
\section{\code{run_corpus_explorer}}{

For \code{run_corpus_explorer}, use \code{\link{explore}}.
}

\keyword{internal}
