% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_data.R
\name{create_test_data}
\alias{create_test_data}
\title{Create test_data}
\usage{
create_test_data()
}
\value{
A corpusexplorationobject used for testing
}
\description{
Create test_data
}
\keyword{internal}
