% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_sotu.R
\name{create_sotu_df}
\alias{create_sotu_df}
\title{Create a data frame with State of the Union texts and metadata}
\usage{
create_sotu_df()
}
\value{
data frame
}
\description{
From the "sotu" package.
}
\keyword{internal}
