% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directly_explore.R
\name{directly_explore}
\alias{directly_explore}
\title{Directly explore data frame or character vector}
\usage{
directly_explore(
  dataset,
  arguments_to_prepare_data = list(use_matrix = FALSE),
  arguments_to_run_corpus_explorer = list()
)
}
\arguments{
\item{dataset}{Data frame or character vector as specified in \code{prepare_data()}}

\item{arguments_to_prepare_data}{List. Arguments to be passed to
\code{prepare_data()} in order to overrule this function's
default argument values.}

\item{arguments_to_run_corpus_explorer}{List. Arguments to be passed to
\code{run_corpus_explorer()} in order to overrule this function's
default argument values.}
}
\value{
Launches a Shiny app.
}
\description{
Function to directly explore data frame or character vector
  without first creating a corporaexplorerobject using
  \code{prepare_data()}, instead creating one on the fly as the app
  launches.
  Functionally equivalent to
  \code{run_corpus_explorer(prepare_data(dataset, use_matrix = FALSE))}.
  Interrupt R to stop the
  application (usually by pressing Ctrl+C or Esc).
}
\details{
By default, no document term matrix will be generated,
  meaning that the data will be prepared for exploration faster than
  by using the default settings in \code{prepare_data()},
  but also that searches in the app are likely to be slower.
}
\examples{
if (interactive()) {

directly_explore(rep(sample(LETTERS), 10))

directly_explore(rep(sample(LETTERS), 10),
  arguments_to_run_corpus_explorer = list(search_input = list(search_terms = "Z"))
)

}
}
