% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore}
\usage{
explore(x, ...)
}
\arguments{
\item{x}{corporaexplorerobject, data.frame or character vector as specified in
\code{run_corpus_explorer()} and \code{directly_explore()}.}

\item{...}{Further arguments to be passed to
\code{run_corpus_explorer()} or \code{directly_explore()}}
}
\value{
Launch Shiny app
}
\description{
Convenience function. Calls \code{run_corpus_explorer()} if \code{x} is a "corporaexplorerobject",
  otherwise it calls \code{directly_explore()}.
  Interrupt R to stop the
  application (usually by pressing Ctrl+C or Esc).
}
\keyword{internal}
