% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_convenience_functions.R
\name{get_df}
\alias{get_df}
\title{Retrieve the document data frame from a corporaexplorerobject}
\usage{
get_df(x, make_normal = TRUE)
}
\arguments{
\item{x}{corporaexplorerobject}
}
\value{
data.frame
}
\description{
Retrieve the document data frame from a corporaexplorerobject
}
\keyword{internal}
