% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_matrix}
\alias{get_matrix}
\title{Split up returned list from matrix_via_r()}
\usage{
get_matrix(returned_list)
}
\arguments{
\item{returned_list}{Returned list from matrix_via_r()}
}
\value{
Document term matrix (data.table).
}
\description{
Split up returned list from matrix_via_r()
}
\keyword{internal}
