% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_term_vector}
\alias{get_term_vector}
\title{Split up returned list from matrix_via_r()}
\usage{
get_term_vector(returned_list)
}
\arguments{
\item{returned_list}{Returned list from matrix_via_r()}
}
\value{
Word vector (character vector).
}
\description{
Split up returned list from matrix_via_r()
}
\keyword{internal}
