% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{is_last_in_month}
\alias{is_last_in_month}
\title{Check if date is the last in the month}
\usage{
is_last_in_month(date)
}
\arguments{
\item{date}{Date.}
}
\value{
Logical.
}
\description{
Check if date is the last in the month
}
\keyword{internal}
