% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickly_explore.R
\name{quickly_explore}
\alias{quickly_explore}
\title{Directly explore data frame or character vector}
\usage{
quickly_explore(
  dataset,
  arguments_to_prepare_data = list(),
  arguments_to_run_corpus_explorer = list()
)
}
\arguments{
\item{dataset}{Data frame or character vector as specified in \code{prepare_data()}}

\item{arguments_to_prepare_data}{List. Further arguments to be passed to
\code{prepare_data()}}

\item{arguments_to_run_corpus_explorer}{List. Further arguments to be passed to
\code{run_corpus_explorer()}}
}
\value{
Launches a Shiny app.
}
\description{
Directly explore data frame or character vector
  without first creating a corporaexplorerobject using
  \code{prepare_data()}.
  Functionally equivalent to
  \code{run_corpus_explorer(prepare_data(dataset))}
}
\examples{
if (interactive()) {

quickly_explore(rep(sample(LETTERS), 10))

quickly_explore(rep(sample(LETTERS), 10),
  arguments_to_run_corpus_explorer = list(search_input = list(search_terms = "Z"))
)

}
}
