% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_corpus_explorer.R
\name{run_corpus_explorer-deprecated}
\alias{run_corpus_explorer-deprecated}
\title{Deprecated: run_corpus_explorer()}
\description{
Deprecated. Use \code{explore()} instead.
}
\seealso{
\code{\link{corporaexplorer-deprecated}}
}
\keyword{internal}
