% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_corpus_explorer.R
\name{run_corpus_explorer}
\alias{run_corpus_explorer}
\title{Launching Shiny app for exploration of text collection}
\usage{
run_corpus_explorer(corpus_object, use_matrix = TRUE,
  regex_engine = c("default", "stringr", "re2r"),
  optional_info = FALSE, ...)
}
\arguments{
\item{corpus_object}{A corpusexploration object created by
\code{\link[corpusexplorationr]{prepare_data}}.}

\item{use_matrix}{Logical. If the corpusexplorationobject contains a document
term matrix, should it be used for searches? (See
\code{\link[corpusexplorationr]{prepare_data}}.)}

\item{regex_engine}{Specify regular expression engine to be used: \itemize{
\item `default`: use the \code{re2r} package
(\url{https://github.com/qinwf/re2r}) for simple searches and the
\code{stringr} package (\url{https://github.com/tidyverse/stringr} for
complex regexes (i.e. when special regex characters are used). \item
`stringr`: use \code{stringr} for all searches. \item `re2r`: use
\code{re2r} for all searches.}}

\item{optional_info}{Logical. If \code{TRUE}, information about search method
(regex engine and whether the search was conducted in the document term
matrix or in the full text documents).}

\item{...}{Other arguments passed to \code{\link[shiny]{runApp}} in the Shiny
package.}
}
\description{
Launch Shiny app for exploration of text collection. Interrupt R to stop the
application (usually by pressing Ctrl+C or Esc).
}
\examples{
# Constructing test data frame:
dates <- as.Date(paste(2011:2020, 1:10, 21:30, sep = "-"))
texts <- paste0(
  "This is a document about ", month.name[1:10], ". ",
  "This is not a document about ", rev(month.name[1:10]), "."
)
titles <- paste("Text", 1:10)
test_df <- tibble::tibble(Date = dates, Text = texts, Title = titles)

# Converting to corpusexploration object:
corpus <- prepare_data(test_df, corpus_name = "Test corpus")
\dontrun{
# Running exploration app:
run_corpus_exploration(corpus)

# Running app to extract documents:
run_document_extractor(corpus)
# Or:
run_document_extractor(test_df)
}
}
