% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_app.R
\name{run_demo_app}
\alias{run_demo_app}
\alias{create_demo_app}
\title{Demo app: Jane Austen's novels}
\usage{
run_demo_app(...)

create_demo_app()
}
\arguments{
\item{...}{Arguments passed to \code{run_corpus_explorer()}}
}
\value{
\code{run_demo_app()} launches a Shiny app. \code{create_demo_app()} returns
  a corporaexplorerobject.
}
\description{
\code{run_demo_app()} is a convenience function to directly
  run the demo app without first creating
  a corporaexplorerobject.
  Equals \code{run_corpus_explorer(create_demo_app)}.
  Interrupt R to stop the
  application (usually by pressing Ctrl+C or Esc).
}
\details{
The demo app's data are Jane Austen's six novels, retrieved
  through the "janeaustenr" package
  (\url{https://github.com/juliasilge/janeaustenr}) --
  which must be installed for these functions to work --
  and converted to a corporaexplorerobject as shown at
  \url{https://kgjerde.github.io/corporaexplorer/articles/jane_austen.html}.
}
\examples{
## Create corporaexplorerobject for demo app:
jane_austen <- create_demo_app()

if(interactive()){

## Run the corporaexplorerobject:
run_corpus_explorer(jane_austen)

## Or create and run the demo app in one step:

run_demo_app()

}
}
