% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_download_tool.R
\name{run_download_tool}
\alias{run_download_tool}
\title{Launching Shiny app for downloading documents from text collection}
\usage{
run_download_tool(data, ...)
}
\arguments{
\item{data}{Either a \code{corpusexploration} object created by
\code{\link[corpusexplorationr]{prepare_data}} or a data frame containing
\code{Date} and \code{Text} columns.}

\item{...}{Other arguments passed to \code{\link[shiny]{runApp}} in the Shiny
package.}
}
\description{
Launch Shiny app for downloading documents from text collection.
Interrupt R to stop the application (usually by pressing Ctrl+C or Esc).
}
\examples{
# Constructing test data frame:
dates <- as.Date(paste(2011:2020, 1:10, 21:30, sep = "-"))
texts <- paste0("This is a document about ", month.name[1:10], ". ",
   "This is not a document about ", rev(month.name[1:10]), ".")
titles <- paste("Text", 1:10)
test_df <- data.frame(Date = dates, Text = texts, Title = titles)

# Converting to corpusexploration object:
corpus <- prepare_data(test_df, corpus_name = "Test corpus")

\dontrun{
# Running exploration app:
run_corpus_exploration(corpus)

# Running download app:
run_download_tool(corpus)
# Or:
run_download_tool(test_df)
}
}
