% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{transform_365}
\alias{transform_365}
\title{Convert "data_dok" tibble to "data_365" tibble}
\usage{
transform_365(new_df)
}
\arguments{
\item{new_df}{A "data_dok" tibble produced by \code{transform_regular}}
}
\value{
A "data_365" tibble.
}
\description{
Convert "data_dok" tibble to "data_365" tibble
}
\keyword{internal}
