% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{transform_regular}
\alias{transform_regular}
\title{Adjusts data frame to corporaexplorer format}
\usage{
transform_regular(df, normalise = TRUE)
}
\arguments{
\item{df}{Data frame with Date column (Date), Text column (character), and
optionally Title (character), URL (character), and Type (character)
columns.}

\item{normalise}{Should non-breaking spaces (U+00A0) and soft hyphens
(U+00ad) be normalised?}
}
\value{
A tibble ("data_dok")
}
\description{
Adjusts data frame to corporaexplorer format
}
\keyword{internal}
