% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_sotu.R
\name{demo_sotu}
\alias{demo_sotu}
\alias{run_sotu_app}
\alias{create_sotu_app}
\alias{run_sotu_decade_app}
\alias{create_sotu_decade_app}
\title{Demo apps: State of the Union addresses}
\usage{
run_sotu_app(...)

create_sotu_app()

run_sotu_decade_app(...)

create_sotu_decade_app()
}
\arguments{
\item{...}{Arguments passed to run_corpus_explorer()}
}
\value{
The \code{run_sotu_*} functions launch a Shiny app.
  The \code{create_sotu_*} functions return a \code{corporaexplorerobject}.
}
\description{
Two demo apps exploring the United States Presidential State of the Union addresses.
  The data are provided by the
  \href{https://cran.r-project.org/web/packages/sotu/index.html}{sotu}
  package, and includes all adresses through 2016.
  Interrupt R to stop the application (usually by pressing Ctrl+C or Esc).
}
\details{
For details, see
  \url{https://kgjerde.github.io/corporaexplorer/articles/sotu.html}.
}
