% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for corpus exploration}
\usage{
prepare_data(dataset, columns_to_include = NULL,
  columns_doc_info = c("Date", "Title", "URL"), corpus_name = NULL,
  use_matrix = TRUE, normalise = TRUE,
  matrix_without_punctuation = TRUE)
}
\arguments{
\item{dataset}{A data frame with Date column (class Date), Text column (class
character), and optionally other columns, e.g. Title (class character) and URL
(character).}

\item{columns_to_include}{Character vector. The columns from df to be
included in corpus object, in addition to the required Date and Text. E.g.
\code{c("Title", "URL")}. If NULL, the default, all columns will be
included.}

\item{columns_doc_info}{Character vector. The columns from df to display in
the "document information" tab in the corpus exploration app. By default
\code{Date}, \code{Title} and \code{URL} will be
displayed, if included. If columns_doc_info includes a column which is not
present in dataset, it will be ignored.}

\item{corpus_name}{Optional character string with name of corpus.}

\item{use_matrix}{Logical. Should the function create a document term matrix
for fast searching? If \code{TRUE}, data preparation will run longer and demand
more memory. If \code{FALSE}, the returning corporaexplorerobject will be more light-weight, but
searching will be slower.}

\item{normalise}{Should non-breaking spaces (U+00A0) and soft hyphens
(U+00ad) be normalised?}

\item{matrix_without_punctuation}{Should punctuation and digits be stripped
  from the text before constructing the document term matrix? If \code{TRUE},
  the default:
\itemize{
    \item The corpusexplorer object will be lighter and most searches in
    the corpus exploration app will be faster.
    \item Searches including punctuation and digits will be carried out in
    the full text documents.
    \item The only "risk" with this strategy is that the corpus exploration
    app in some cases can produce false positives. E.g. searching for the
    term "donkey" will also find the term "don\%key".
This should not be a problem for the vast opportunity of use cases, but if
one so desires, there are three different solutions: set this paramater to
\code{FALSE}, create a corporaexplorerobject without a matrix by setting
the \code{use_matrix} parameter to \code{FALSE}, or run
\code{\link[corporaexplorer]{run_corpus_explorer}} with the
\code{use_matrix} parameter set to \code{FALSE}.
}
 If \code{FALSE}, the corpusexplorer object will be larger, and most
 simple searches will be slower.}
}
\value{
A \code{corpusexplorer} object to be passed as argument to
  \code{\link[corporaexplorer]{run_corpus_explorer}} and
  \code{\link[corporaexplorer]{run_document_extractor}}.
}
\description{
Prepare data for corpus exploration
}
\examples{
# Constructing test data frame:
dates <- as.Date(paste(2011:2020, 1:10, 21:30, sep = "-"))
texts <- paste0(
  "This is a document about ", month.name[1:10], ". ",
  "This is not a document about ", rev(month.name[1:10]), "."
)
titles <- paste("Text", 1:10)
test_df <- tibble::tibble(Date = dates, Text = texts, Title = titles)

# Converting to corpusexplorer object:
corpus <- prepare_data(test_df, corpus_name = "Test corpus")
\dontrun{
# Running exploration app:
run_corpus_exploration(corpus)

# Running app to extract documents:
run_document_extractor(corpus)
}
}
