% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{print.corporaexplorerobject}
\alias{print.corporaexplorerobject}
\title{Print corporaexplorerobject}
\usage{
\method{print}{corporaexplorerobject}(x, ...)
}
\arguments{
\item{x}{A corporaexplorerobject}
}
\value{
Console-friendly output
}
\description{
Print corporaexplorerobject
}
\keyword{internal}
