% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{A tiny test dataset to test basic functionality}
\format{A corpusexplorationobject.}
\usage{
test_data
}
\description{
Created by \code{corpusexplorationr:::create_test_data()}.
}
\keyword{datasets}
