% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{transform_regular}
\alias{transform_regular}
\title{Adjusts data frame to corporaexplorer format}
\usage{
transform_regular(df)
}
\arguments{
\item{df}{Data frame with Date column (Date), Text column (character), and
optionally Title (character), URL (character), and Type (character)
columns.}
}
\value{
A tibble ("data_dok")
}
\description{
Adjusts data frame to corporaexplorer format
}
\keyword{internal}
